// src/types/enums.ts
export enum grade_e { sdt = 'sdt', cpl = 'cpl', cch = 'cch', cc1 = 'cc1', sgt = 'sgt', sch = 'sch', adj = 'adj', adc = 'adc', maj = 'maj', ltn = 'ltn', cne = 'cne' }

export enum services_e {
  chef_ps = 'chef_ps',
  planton_ps = 'planton_ps',
  chef_spi = 'chef_spi',
  cpa = 'cpa', cp = 'cp',
  semaine = 'semaine',
  planton_ei = 'planton_ei',
  chef_ei = 'chef_ei'
}